Muchos cambios.
1) OPTIMIZACION Y LIMPIEZA DE CODIGO
 - Uso de co rutinas en unity, optimiza bastante los bucles y hace mas facil programar algunas cosas, tengo que seguir substituyendo bucles por co rutinas.
- Gracias a las co rutinas, comprobaciones que se hacian cada frame de forma innecesaria, ahora se comprueban en intervalos fijados de tiempo.
- Mejorado de forma masiva el rendimiento del interfaz, no se actualiza cada frame si no es necesario, y cambiados algunos ajustes para evitar perdidas de FPS sin ningun cambio que se pueda notar en el aspecto visual.

2) FIXES

- Arreglado el problema con la iluminacion, cada vez que el jugador se desplaza 1000 unidades en Z, o X, se mueve un parent que contiene la escena 1000 unidades hacia atras en esas coordenadas, 0 impacto en rendimiento.

3) GAMEMODES

- Añadida la base para diferentes modos de juego, deben seleccionarse en StageData.Gamemode, los modos son los siguientes:
0 - Freeroam: Sin restricciones ni tiempo, para probar simplemente.
1 - Endurance: Lo que teniamos hecho hasta ahora.
2 - Drift Endurance: Lo mismo, pero la unica forma de conseguir tiempo extra es con drifts.
3 - Drift Exhibition: Solo se consigue tiempo extra al cruzar puntos de control, solo cuenta la puntuacion de drift obtenida. Hay un multiplicador de drift que aumenta en 0.1 cada 2 segundos de derrape mantenido.
4: High Speed Challenge: Como endurance, pero todo el tiempo obtenido se reduce un 90%, el tiempo se congela si estas conduciendo por encima del 90% de la velocidad maxima del coche. Daños recibidos aumentados un 50%.
5: Chain Drift Challenge: Como endurance, pero solo se consigue tiempo extra de los puntos de control, y estos dan un 95% menos de tiempo (apenas aumentan el tiempo 0.5 segundos), tienes solo 5 segundos de tiempo, pero el tiempo se congela mientras derrapas.
6 - Time attack: Pensado para contrarelojes, aun no lo tengo preparado, la idea es que el modo este basado en llegar a X punto de control en menos de Y tiempo.

EN PROCESO:

- Cambios de interfaz, mover el tiempo al centro quizas?
- Mejorar visualmente las notificaciones...me falta un poco.
- Pantalla de final de prueba especifica del GameMode, plantilla preparada, me falta escribir la informacion relevante de cada modo.
- Pantalla de presentacion de prueba, una camara girando alrededor del coche, con un efecto de desenfoque, y en primer plano un panel con la informacion del evento, y requisitos para oro, plata, bronce (si hay), o laderboards, si es una prueba que compitas contra otros jugadores.

+++ Assets/RoadGenerator.cs
@@ -88,7 +88,7 @@ public class RoadGenerator : MonoBehaviour {
 		lastReadedNode.SetLighScale (globalRoadScale);
 		//Aqui creamos CheckPoint
 		if (nodesSinceLastActiveCP >= nodesBetweenActiveCP) {
-			lastReadedNode.SetAsActiveCheckpoint ((int)(stackedNodeWeight * NodeWeight2Time));
+			lastReadedNode.SetAsActiveCheckpoint ((stackedNodeWeight * NodeWeight2Time));
 			nodesSinceLastActiveCP = 0;
 			stackedNodeWeight = 0;
 			NodeWeight2Time *= 0.98f;
